[core] Move FirehoseClient to shared folder
