+++ streamalert/classifier/clients/__init__.py
@@ -1,3 +1,2 @@
 """Imports from submodules to make higher up imports easier"""
-from .firehose import FirehoseClient
 from .sqs import SQSClient
