+++ streamalert/shared/firehose.py
