+++ streamalert_cli/athena/handler.py
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-from streamalert.classifier.clients import FirehoseClient
+from streamalert.shared.firehose import FirehoseClient
 from streamalert.shared.utils import get_database_name, get_data_file_format
 from streamalert.shared.alert import Alert
 from streamalert.shared.athena import AthenaClient
